Moved utils module to shoreline.
